Get and create order

+++ Backend/StockWise/Dtos/CreateOrderDto.cs
@@ -3,8 +3,8 @@
     public class CreateOrderDto
     {
         public string SellerName {  get; set; }
-        public int BuyerNIP {  get; set; }
-        public int SellerNIP { get; set; }
+        public long BuyerNIP {  get; set; }
+        public long SellerNIP { get; set; }
         public string Address { get; set; }
         public string? Email { get; set; }
         public string? Phone { get; set; }
@@ -12,6 +12,7 @@
 
         public List<String> ProductsEAN {  get; set; }
         public int Quantity { get; set; } = 0;
+
 
     }
 }
